working version with optimized database setup

+++ lib/sql_service.dart
@@ -12,6 +12,7 @@ import 'models/entitlement.dart';
 import 'models/question.dart';
 
 class SqliteService with ChangeNotifier {
+  static const NEW_DB_VERSION = 2;
   SqliteService() {
     _initDatabase();
   }
@@ -22,12 +23,15 @@ class SqliteService with ChangeNotifier {
   static List<Question> _quesionListTemp = [];
   static bool _flag = false;
   Question randQuestion =
-      new Question(german: "Laden...", english: "Loasing", type: "free");
+      new Question(german: "Laden...", english: "Loading", type: "free");
   Future<Database> get database async => _database ??= await _initDatabase();
 
   Future<Database> _initDatabase() async {
+    Database db;
     var databasesPath = await getDatabasesPath();
     String path = join(databasesPath, 'question_database.db');
+
+    // if db doesn't exist at all
     if (FileSystemEntity.typeSync(path) == FileSystemEntityType.notFound) {
       //Load db from assets folder
       ByteData data =
@@ -35,15 +39,29 @@ class SqliteService with ChangeNotifier {
       List<int> bytes =
           data.buffer.asUint8List(data.offsetInBytes, data.lengthInBytes);
       await File(path).writeAsBytes(bytes);
+      db = await openDatabase(path);
+      db.setVersion(NEW_DB_VERSION);
     } else {
-      deleteDatabase(path);
-      ByteData data =
-          await rootBundle.load(join('assets', 'questions_database.db'));
-      List<int> bytes =
-          data.buffer.asUint8List(data.offsetInBytes, data.lengthInBytes);
-      await File(path).writeAsBytes(bytes);
+      db = await openDatabase(path);
+
+      // if database exists but version is to low
+      if (await db.getVersion() < NEW_DB_VERSION) {
+        deleteDatabase(path);
+        ByteData data =
+            await rootBundle.load(join('assets', 'questions_database.db'));
+        List<int> bytes =
+            data.buffer.asUint8List(data.offsetInBytes, data.lengthInBytes);
+        await File(path).writeAsBytes(bytes);
+        db = await openDatabase(path);
+        db.setVersion(NEW_DB_VERSION);
+      }
+
+      // database exists and version is ok
+      else {
+        db = await openDatabase(path);
+      }
     }
-    return openDatabase(path);
+    return db;
   }
 
   //Delete Database
